requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.0
ollama>=0.1.8
faiss-cpu>=1.8.0
numpy>=1.26.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
python-telegram-bot>=21.0
aiosmtplib>=3.0.1
jinja2>=3.1.3
pytest>=8.1.0
pytest-asyncio>=0.23.0
loguru>=0.7.2
tenacity>=8.2.3
aiosqlite>=0.20.0
markdown>=3.5.0
tqdm>=4.66.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
